backend/requirements.txt
# Backend Python dependencies

fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.7.1
scikit-learn==1.4.2
numpy==1.26.4
python-dotenv==1.0.1
httpx==0.27.0
anthropic>=0.25.0
redis>=5.0.4
